Move `enum` definitions to `libOPHD`
Search replace for header includes:
`#include "(.*)Enum(.*).h"`
`#include <libOPHD/Enum$2.h>`

Followed by manually moving the includes to the appropriate section.
