+++ appOPHD/GraphWalker.cpp
@@ -1,10 +1,11 @@
-#include "EnumDirection.h"
 #include "GraphWalker.h"
 #include "DirectionOffset.h"
 
 #include "Map/MapCoordinate.h"
 #include "Map/TileMap.h"
 #include "MapObjects/Structure.h"
+
+#include <libOPHD/EnumDirection.h>
 
 
 using namespace NAS2D;
