+++ appOPHD/States/CrimeExecution.cpp
@@ -1,10 +1,10 @@
 #include "CrimeExecution.h"
 
 #include "../Resources.h"
-#include "../EnumDifficulty.h"
 #include "../StructureManager.h"
 #include "../MapObjects/Structures/FoodProduction.h"
 
+#include <libOPHD/EnumDifficulty.h>
 #include <libOPHD/RandomNumberGenerator.h>
 
 #include <NAS2D/StringUtils.h>
