+++ appOPHD/States/MapViewStateHelper.cpp
@@ -13,13 +13,14 @@
 #include "../Constants/Strings.h"
 #include "../StructureCatalogue.h"
 #include "../StructureManager.h"
-#include "../EnumDirection.h"
 #include "../DirectionOffset.h"
 #include "../Map/TileMap.h"
 #include "../MapObjects/StructureType.h"
 #include "../MapObjects/Structures/Warehouse.h"
 
 #include "../UI/MessageBox.h"
+
+#include <libOPHD/EnumDirection.h>
 
 #include <NAS2D/Utility.h>
