+++ appOPHD/States/MapViewStateTurn.cpp
@@ -11,10 +11,11 @@
 
 #include "../Cache.h"
 #include "../DirectionOffset.h"
-#include "../EnumMoraleIndex.h"
 #include "../MoraleString.h"
 #include "../StorableResources.h"
 #include "../StructureManager.h"
+
+#include <libOPHD/EnumMoraleIndex.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
