+++ appOPHD/States/MapViewStateUi.cpp
@@ -12,7 +12,6 @@
 #include "../Constants/Strings.h"
 #include "../Constants/UiConstants.h"
 
-#include "../EnumDirection.h"
 #include "../DirectionOffset.h"
 #include "../PointerType.h"
 #include "../StructureCatalogue.h"
@@ -24,6 +23,8 @@
 #include "../UI/MessageBox.h"
 #include "../UI/NavControl.h"
 #include "../UI/CheatMenu.h"
+
+#include <libOPHD/EnumDirection.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
