+++ appOPHD/UI/Reports/FactoryReport.h
@@ -2,11 +2,13 @@
 
 #include "ReportInterface.h"
 
+#include "../FactoryListBox.h"
+
+#include <libOPHD/EnumProductType.h>
+
 #include <libControls/Button.h>
 #include <libControls/ComboBox.h>
 #include <libControls/TextArea.h>
-#include "../FactoryListBox.h"
-#include "../../EnumProductType.h"
 
 #include <NAS2D/Math/Rectangle.h>
