+++ appOPHD/appOPHD.vcxproj.filters
@@ -359,36 +359,6 @@
     <ClInclude Include="DirectionOffset.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="EnumConnectorDir.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumDifficulty.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumDirection.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumDisabledReason.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumIdleReason.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumMineProductionRate.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumMoraleIndex.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumProductType.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumStructureID.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="EnumTerrainType.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="GraphWalker.h">
       <Filter>Header Files</Filter>
     </ClInclude>
