+++ libOPHD/EnumDirection.h
