+++ libOPHD/EnumIdleReason.h
