+++ libOPHD/EnumMineProductionRate.h
