+++ libOPHD/EnumMoraleIndex.h
