+++ libOPHD/libOPHD.vcxproj
@@ -172,6 +172,16 @@
     <ClCompile Include="XmlSerializer.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="EnumConnectorDir.h" />
+    <ClInclude Include="EnumDifficulty.h" />
+    <ClInclude Include="EnumDirection.h" />
+    <ClInclude Include="EnumDisabledReason.h" />
+    <ClInclude Include="EnumIdleReason.h" />
+    <ClInclude Include="EnumMineProductionRate.h" />
+    <ClInclude Include="EnumMoraleIndex.h" />
+    <ClInclude Include="EnumProductType.h" />
+    <ClInclude Include="EnumStructureID.h" />
+    <ClInclude Include="EnumTerrainType.h" />
     <ClInclude Include="Map\MapOffset.h" />
     <ClInclude Include="RandomNumberGenerator.h" />
     <ClInclude Include="Population\Population.h" />
